+++ tests/test_graphhopper.py
@@ -20,6 +20,7 @@ from routingpy import Graphhopper
 from routingpy.direction import Direction, Directions
 from routingpy.isochrone import Isochrones, Isochrone
 from routingpy.matrix import Matrix
+from routingpy.utils import decode_polyline5
 
 from tests.test_helper import *
 import tests as _test
@@ -91,7 +92,6 @@ class GraphhopperTest(_test.TestCase):
             "&points_encoded=true&vehicle=car&type=json&weighting=short_fastest&snap_prevention=trunk%2Cferry&curb_side=any%2Cright&turn_costs=true",
             responses.calls[0].request.url,
         )
-
         self.assertIsInstance(routes, Directions)
         self.assertEqual(3, len(routes))
         self.assertIsInstance(routes[0], Direction)
@@ -99,6 +99,41 @@ class GraphhopperTest(_test.TestCase):
         self.assertIsInstance(routes[0].duration, int)
         self.assertIsInstance(routes[0].distance, int)
         self.assertIsInstance(routes[0].raw, dict)
+
+    @responses.activate
+    def test_full_directions_not_encoded(self):
+        query = deepcopy(ENDPOINTS_QUERIES[self.name]["directions"])
+        query["points_encoded"] = False
+        query["algorithm"] = None
+
+        res = ENDPOINTS_RESPONSES[self.name]["directions"]
+        res["paths"][0]["points"] = dict(coordinates=decode_polyline5(res["paths"][0]["points"]))
+
+        responses.add(
+            responses.GET,
+            "https://graphhopper.com/api/1/route",
+            status=200,
+            json=ENDPOINTS_RESPONSES[self.name]["directions"],
+            content_type="application/json",
+        )
+
+        route = self.client.directions(**query)
+        self.assertEqual(1, len(responses.calls))
+        self.assertURLEqual(
+            "https://graphhopper.com/api/1/route?avoid=tunnel%3Bford&"
+            "block_area=48.23424%2C8.34234&calc_points=false&ch.disable=true&debug=true&details=time&details=tolls&"
+            "elevation=true&heading=50%2C50%2C50&heading_penalty=100&instructions=false&key=sample_key&locale=en&"
+            "optimize=true&pass_through=true&point=49.415776%2C8.680916&point=49.420577%2C8.688641&"
+            "point=49.445776%2C8.780916&point_hint=Graphhopper%20Lane&point_hint=OSM%20Street&point_hint=Routing%20Broadway"
+            "&points_encoded=false&vehicle=car&type=json&weighting=short_fastest&snap_prevention=trunk%2Cferry&curb_side=any%2Cright&turn_costs=true",
+            responses.calls[0].request.url,
+        )
+
+        self.assertIsInstance(route, Direction)
+        self.assertIsInstance(route.geometry, list)
+        self.assertIsInstance(route.duration, int)
+        self.assertIsInstance(route.distance, int)
+        self.assertIsInstance(route.raw, dict)
 
     @responses.activate
     def test_full_isochrones(self):